Extract common document capture CSP handling to concern
See: https://github.com/18F/identity-idp/pull/4675#discussion_r577842923
