+++ app/controllers/idv/doc_auth_controller.rb
@@ -4,10 +4,10 @@ module Idv
     before_action :redirect_if_mail_bounced
     before_action :redirect_if_pending_profile
     before_action :extend_timeout_using_meta_refresh_for_select_paths
-    before_action :override_document_capture_step_csp
 
     include IdvSession # remove if we retire the non docauth LOA3 flow
     include Flow::FlowStateMachine
+    include Idv::DocumentCaptureConcern
 
     before_action :update_if_skipping_upload
 
@@ -48,18 +48,6 @@ module Idv
 
     def flow_session
       user_session['idv/doc_auth']
-    end
-
-    def override_document_capture_step_csp
-      return unless params[:step] == 'document_capture'
-
-      SecureHeaders.append_content_security_policy_directives(
-        request,
-        # required to run wasm until wasm-eval is available
-        script_src: ['\'unsafe-eval\''],
-        # required for retrieving image dimensions from uploaded images
-        img_src: ['blob:'],
-      )
     end
   end
 end
